agentic-system/pyproject.toml
[project]
name = "agentic-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Stijn Janssens",email = "janssens.stijn@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "pydantic-ai (>=0.0.38,<0.0.39)",
    "app @ file:///Users/stijn/Documents/data-minded/projects/data-product-portal/backend",
    "logfire[asyncpg,httpx] (>=3.6.4,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
